+++ packages/zowe-explorer-api/src/fs/BaseProvider.ts
@@ -488,25 +488,4 @@ export class BaseProvider {
             silent ?? false
         );
     }
-    protected async awaitExtenderType(
-        profileInfo: any, //Profiles
-        profileName: string,
-        getCallback: (getName: string) => any,
-        setCallback: (setName: string, setPromise: DeferredPromise<void>) => void
-    ): Promise<void> {
-        const profile = profileInfo.allProfiles.find((prof) => prof.name === profileName);
-        if (profile && profile.type !== "zosmf" && !getCallback(profileName)) {
-            const deferredPromise = new DeferredPromise<void>();
-            setCallback(profileName, deferredPromise);
-        }
-        const profilePromise = getCallback(profileName);
-        const promiseTimeout = 10000;
-        if (profilePromise) {
-            let timeoutHandle: NodeJS.Timeout;
-            const timeoutPromise = new Promise<void>((_, reject) => {
-                timeoutHandle = setTimeout(() => reject(new Error("Timeout waiting for profile")), promiseTimeout);
-            });
-            await Promise.race([profilePromise.promise.finally(() => clearTimeout(timeoutHandle)), timeoutPromise]);
-        }
-    }
 }
