+++ packages/zowe-explorer/src/trees/dataset/DatasetFSProvider.ts
@@ -512,15 +512,22 @@ export class DatasetFSProvider extends BaseProvider implements vscode.FileSystem
         // If the extenderTypeReady map does not contain the profile, create a deferred promise for the profile.
         const profileName = uri.path.split("/")[1];
         const profileInfo = Profiles.getInstance();
+        const profile = profileInfo.allProfiles.find((prof) => prof.name === profileName);
+        if (profile && profile.type !== "zosmf" && !Profiles.extenderTypeReady.get(profileName)) {
+            const deferredPromise = new DeferredPromise<void>();
+            Profiles.extenderTypeReady.set(profileName, deferredPromise);
+        }
 
-        const getCallback = (getName: string): any => {
-            return Profiles.extenderTypeReady.get(getName);
-        };
-        const setCallback = (setName: string, setPromise: DeferredPromise<void>): void => {
-            Profiles.extenderTypeReady.set(setName, setPromise);
-        };
+        const profilePromise = Profiles.extenderTypeReady.get(profileName);
+        const promiseTimeout = 10000;
+        if (profilePromise) {
+            let timeoutHandle: NodeJS.Timeout;
+            const timeoutPromise = new Promise<void>((resolve, _) => {
+                timeoutHandle = setTimeout(() => resolve(), promiseTimeout);
+            });
 
-        await this.awaitExtenderType(profileInfo, profileName, getCallback, setCallback);
+            await Promise.race([profilePromise.promise.finally(() => clearTimeout(timeoutHandle)), timeoutPromise]);
+        }
 
         try {
             ds = this._lookupAsFile(uri) as DsEntry;
